libft

+++ ft_arrsize.c
@@ -1,14 +1,26 @@
-#include "./includes/libft.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_arrsize.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: efichot <efichot@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2016/11/09 11:06:40 by efichot           #+#    #+#             */
+/*   Updated: 2016/11/09 12:26:12 by efichot          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-int ft_arrsize(char **arr)
+#include "libft.h"
+
+size_t	ft_arrsize(char **arr)
 {
-  int i;
+	size_t	i;
 
-  i = 0;
-  if (arr)
-  {
-    while (arr[i])
-      i++;
-  }
-  return (i);
+	i = 0;
+	if (arr)
+	{
+		while (arr[i])
+			i++;
+	}
+	return (i);
 }
